removing dependence on BaseReplicationExecuter and BaseReplicationLoader
